Fetching the data from the server via HTTP
