+++ Advance_Redux/src/store/cart-actions.js
@@ -1,0 +1,93 @@
+import { uiActions } from "./ui-slice";
+import { cartActions } from "./cart-slice";
+
+export const fetchCartData = () => {
+  return async (dispatch) => {
+    const fetchData = async () => {
+      const response = await fetch(
+        "https://react-redux-http-67823-default-rtdb.asia-southeast1.firebasedatabase.app/cart.json"
+      );
+
+      if (!response.ok) {
+        throw new Error("Could not fetch cart data!");
+      }
+
+      const data = await response.json();
+      return data;
+    };
+
+    try {
+      const cartData = await fetchData();
+      dispatch(cartActions.replaceCart(cartData));
+    } catch (error) {
+      dispatch(
+        uiActions.showNotification({
+          status: "error",
+          title: "Error!",
+          message: "Sending cart data failed!",
+        })
+      );
+    }
+  };
+};
+
+export const sendCartData = (cart) => {
+  return async (dispatch) => {
+    dispatch(
+      uiActions.showNotification({
+        status: "pending",
+        title: "Sending...",
+        message: "Sending cart data!",
+      })
+    );
+
+    const sendRequest = async () => {
+      const response = await fetch(
+        "https://react-redux-http-67823-default-rtdb.asia-southeast1.firebasedatabase.app/cart.json",
+        {
+          method: "PUT",
+          body: JSON.stringify(cart),
+        }
+      );
+
+      if (!response.ok) {
+        throw new Error("Sending cart data failed!");
+      }
+    };
+
+    try {
+      await sendRequest();
+      dispatch(
+        uiActions.showNotification({
+          status: "success",
+          title: "Success!",
+          message: "Sending cart data successfully!",
+        })
+      );
+    } catch (error) {
+      dispatch(
+        uiActions.showNotification({
+          status: "error",
+          title: "Error!",
+          message: "Sending cart data failed!",
+        })
+      );
+    }
+  };
+};
+
+/*
+ dispatch(sendCartData(cart)); - in App.js file
+
+  This might look weird what we dispatch before always were an action creator so function with a type and so on.
+  Now with cart-slice we are instead dispatching a function that returns an another function but the
+  great thing about redux when we are using redux-toolkit is that it is prepared for that so
+  it does not only accepts action object with a type property instead it does accepts action creators which is 
+  a function `sendCartData()` and it will execute that function for us and it will make the dispatch
+  argument property available for us so that code inside of that can make use of it and perform side-effects.
+*/
+
+/* 
+  Using of Action Creator thunk pattern is much more common than the earlier one because it makes a component
+  leaner and thinner i.e., not keeping too much logic inside of the component.
+*/
